Extract home functionality
Extract org.platformio.eclipse.ide.home.core bundle
