+++ bundles/org.platformio.eclipse.ide.home.core/META-INF/MANIFEST.MF
@@ -1,0 +1,14 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: %Bundle-Name
+Bundle-SymbolicName: org.platformio.eclipse.ide.home.core
+Bundle-Version: 0.1.0.qualifier
+Bundle-Vendor: %Bundle-Vendor
+Automatic-Module-Name: org.platformio.eclipse.ide.home.core
+Bundle-RequiredExecutionEnvironment: JavaSE-11
+Require-Bundle: org.eclipse.core.runtime;bundle-version="3.20.100",
+ org.eclipse.osgi.services;bundle-version="3.10.0",
+ org.platformio.eclipse.ide.home;bundle-version="0.1.0",
+ org.platformio.eclipse.ide.installer;bundle-version="0.1.0"
+Bundle-ActivationPolicy: lazy
+Service-Component: OSGI-INF/org.platformio.eclipse.ide.home.internal.core.HomeAccess.xml
