+++ bundles/org.platformio.eclipse.ide.home.core/OSGI-INF/l10n/bundle.properties
@@ -1,4 +1,4 @@
-#Properties file for org.platformio.eclipse.ide.home.ui
+#Properties file for org.platformio.eclipse.ide.home.core
 ###############################################################################
 # Copyright (c) 2021 PlatformIO and ArSysOp.
 #
@@ -21,7 +21,4 @@
 ###############################################################################
 
 Bundle-Vendor = PlatformIO
-Bundle-Name = PlatformIO Eclipse IDE Home UI
-
-perspective.home.name = PlatformIO Home
-view.home.name = PlatformIO Home
+Bundle-Name = PlatformIO Eclipse IDE Home Core
