+++ bundles/org.platformio.eclipse.ide.home.core/OSGI-INF/org.platformio.eclipse.ide.home.internal.core.HomeAccess.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.platformio.eclipse.ide.home.internal.core.HomeAccess">
+   <implementation class="org.platformio.eclipse.ide.home.internal.core.HomeAccess"/>
+</scr:component>
