+++ bundles/org.platformio.eclipse.ide.home.core/build.properties
@@ -1,4 +1,3 @@
-#Properties file for org.platformio.eclipse.ide.home.ui
 ###############################################################################
 # Copyright (c) 2021 PlatformIO and ArSysOp.
 #
@@ -20,8 +19,8 @@
 #     ArSysOp - initial API and implementation
 ###############################################################################
 
-Bundle-Vendor = PlatformIO
-Bundle-Name = PlatformIO Eclipse IDE Home UI
-
-perspective.home.name = PlatformIO Home
-view.home.name = PlatformIO Home
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .,\
+               OSGI-INF/
