+++ bundles/org.platformio.eclipse.ide.home.core/src/org/platformio/eclipse/ide/home/internal/core/messages.properties
@@ -19,6 +19,7 @@
 #     Nikifor Fedorov (ArSysOp) - initial API and implementation
 ###############################################################################
 PlatformIO_installation_message=Installing PlatformIO
+PlatformIO_stop_message=Stopping PlatformIO
 Core_installation_message=Creating virtual environment
 Virtualenv_creation_message=Installing PlatfomIO Core
 Installation_successful_message=Installation completed successfully!
