+++ bundles/org.platformio.eclipse.ide.home.ui/META-INF/MANIFEST.MF
@@ -9,4 +9,6 @@ Bundle-RequiredExecutionEnvironment: JavaSE-11
 Require-Bundle: org.eclipse.core.runtime;bundle-version="3.20.100",
  org.eclipse.ui;bundle-version="3.119.0",
  org.eclipse.ui.workbench;bundle-version="3.122.100",
+ org.platformio.eclipse.ide.home.core;bundle-version="0.1.0",
  org.platformio.eclipse.ide.ui;bundle-version="0.1.0"
+Bundle-ActivationPolicy: lazy
