+++ bundles/org.platformio.eclipse.ide.ui/META-INF/MANIFEST.MF
@@ -8,12 +8,5 @@ Automatic-Module-Name: org.platformio.eclipse.ide.ui
 Bundle-Activator: org.platformio.eclipse.ide.internal.ui.Activator
 Bundle-RequiredExecutionEnvironment: JavaSE-11
 Require-Bundle: org.eclipse.core.runtime;bundle-version="0.0.0",
- org.eclipse.e4.core.contexts;bundle-version="1.8.400",
- org.eclipse.ui.browser;bundle-version="0.0.0",
- org.eclipse.ui.navigator.resources;bundle-version="0.0.0",
- org.eclipse.ui.workbench;bundle-version="0.0.0",
- org.eclipse.ui;bundle-version="0.0.0",
- org.platformio.eclipse.ide.installer;bundle-version="0.0.0",
- org.platformio.eclipse.ide.home;bundle-version="0.0.0",
- org.eclipse.ui.ide;bundle-version="3.18.100"
+ org.eclipse.ui;bundle-version="0.0.0"
 Bundle-ActivationPolicy: lazy
