+++ features/org.platformio.eclipse.ide.feature/feature.xml
@@ -97,6 +97,13 @@
          unpack="false"/>
 
 	<plugin
+         id="org.platformio.eclipse.ide.home.core"
+         download-size="0"
+         install-size="0"
+         version="0.0.0"
+         unpack="false"/>
+
+	<plugin
          id="org.platformio.eclipse.ide.home.ui"
          download-size="0"
          install-size="0"
