+++ releng/org.platformio.eclipse.ide.aggregator/pom.xml
@@ -40,6 +40,7 @@
 		<module>..\..\bundles\org.platformio.eclipse.ide.installer.macosx</module>
 		<module>..\..\bundles\org.platformio.eclipse.ide.installer.linux</module>
 		<module>..\..\bundles\org.platformio.eclipse.ide.home</module>
+		<module>..\..\bundles\org.platformio.eclipse.ide.home.core</module>
 		<module>..\..\bundles\org.platformio.eclipse.ide.home.ui</module>
 		<module>..\..\bundles\org.platformio.eclipse.ide.ui</module>
 		<module>..\..\features\org.platformio.eclipse.ide.feature</module>
